requirements.txt
selenium==4.22.0
webdriver-manager==4.0.2
undetected-chromedriver==3.5.5

requests==2.32.3
requests-oauthlib==2.0.0

setuptools==70.3.0